New strategy: Have a helper program deal w/ the downloading of documents instead of threads.
