+++ .gitignore
@@ -39,6 +39,7 @@ stemwords
 test_chop_into_words
 test_find_substring
 test_auto_temp_file
+update_full_text_db
 
 bsz_daten
 KeywordChainParser-*.jar
