+++ cpp/Makefile
@@ -3,7 +3,7 @@ PROGS=marc_grep jop_grep download_test add_isbns_or_issns_to_articles marc_grep_
       bib_ref_to_codes_tool extract_text_from_image_only_pdfs shared_buffer_test delete_ids marc_filter \
       exec_test krimdok_filter resolve_in_path_test enrich_ddcs control_number_filter stemmer_test \
       enrich_keywords_with_title_words test_find_substring test_chop_into_words test_auto_temp_file \
-      ixtheo_notation_stats
+      ixtheo_notation_stats update_full_text_db
 INSTALL_PROGS=$(filter-out %_test,$(PROGS))
 CGI_PROGS=full_text_lookup
 CCC=g++
@@ -82,13 +82,18 @@ download_test.o: download_test.cc lib/SmartDownloader.h lib/util.h lib/FileUtil.
 download_test: download_test.o lib/libmarc.a
 	$(CCC) -o $@ $^ -Llib -lmarc -lpcre -lcrypto -lkyotocabinet -lmagic
 
-create_full_text_db.o: create_full_text_db.cc lib/Downloader.h lib/RegexMatcher.h lib/StringUtil.h \
-                       lib/TextUtil.h lib/util.h lib/MediaTypeUtil.h lib/MarcUtil.h lib/Subfields.h \
-                       lib/SharedBuffer.h lib/SmartDownloader.h lib/PdfUtil.h lib/ExecUtil.h lib/FileUtil.h \
-	               lib/TimeLimit.h lib/util.h lib/ThreadManager.h
+create_full_text_db.o: create_full_text_db.cc lib/Downloader.h lib/RegexMatcher.h lib/StringUtil.h lib/util.h \
+                       lib/MarcUtil.h lib/Subfields.h lib/ExecUtil.h lib/FileLocker.h lib/TimeLimit.h lib/util.h
 
 create_full_text_db: create_full_text_db.o lib/libmarc.a
 	$(CCC) -o $@ $^ -Llib -lmarc -lpcre -lcrypto -lkyotocabinet -lmagic -lpthread
+
+update_full_text_db: update_full_text_db.o lib/libmarc.a
+	$(CCC) -o $@ $^ -Llib -lmarc -lpcre -lcrypto -lkyotocabinet -lmagic -lpthread
+
+update_full_text_db.o: update_full_text_db.cc lib/DirectoryEntry.h lib/ExecUtil.h lib/FileLocker.h lib/FileUtil.h \
+                       lib/MarcUtil.h lib/MediaTypeUtil.h lib/PdfUtil.h lib/SmartDownloader.h lib/StringUtil.h \
+                       lib/util.h
 
 delete_ids.o: delete_ids.cc lib/MarcUtil.h lib/StringUtil.h lib/Subfields.h lib/util.h
